Fix work order management test
còn file postman

+++ src/main/java/com/ev/warranty/model/dto/claim/ClaimDiagnosticRequest.java
@@ -19,7 +19,7 @@ public class ClaimDiagnosticRequest {
 
     private String repairNotes;
 
-    private Double laborHours;
+    private BigDecimal laborHours; // Changed from Double to BigDecimal for precision
 
     // 🆕 Warranty cost estimate
     private BigDecimal warrantyCost;
